+++ test/cli/cli_exclude_test.dart
@@ -27,6 +27,7 @@ void main() {
       File(p.join(tempDir.path, 'crap4dart.yaml')).writeAsStringSync('''
 coverage:
   required: false
+  run_tests: false
 exclude:
   - 'lib/gen/**'
 ''');